Added context manager for file object
With statement (context manager) automatically closes the file when code within context is finished. File objects should be closed explicitly, especially if you open many simultaneously. Relying on garbage collection is risky when not all installations of python come with one that deals with files, and even then you have little control of when the file handles are freed up by garbage collection. Also easier to debug if error occurs when attempting file opening. See https://stackoverflow.com/questions/1834556/does-a-file-object-automatically-close-when-its-reference-count-hits-zero